+++ src/services/auth/serverActions.ts
@@ -7,7 +7,7 @@ import {
 import { serverDb } from "@/lib/serverDb";
 import { userRole } from "@/types/user";
 
-async function isAdmin(userEmail: string): Promise<boolean> {
+async function isAdminEmail(userEmail: string): Promise<boolean> {
   const adminEmails = process.env.ADMIN_USER_EMAILS;
 
   if (!adminEmails) {
@@ -45,7 +45,7 @@ export async function createProfile(
     return;
   }
 
-  const isAdminUser: boolean = await isAdmin(userEmail);
+  const isAdminUser: boolean = await isAdminEmail(userEmail);
 
   const role: userRole = isAdminUser ? "admin" : "viewer";
